#102 - Control VLC with mididingsd
